Something's wrong with the board's move deduction: be sure to check

+++ src/game/traits/u64_shift.rs
@@ -1,11 +1,11 @@
 use crate::game::model::player::Player;
 
 pub(crate) trait U64Ext {
-    fn shift_by(&self, shift: u8, player: Player) -> u64;
+    fn shift_by(&self, shift: u8, player: Player) -> u32;
 }
 
-impl U64Ext for u64 {
-    fn shift_by(&self, shift: u8, player: Player) -> u64 {
+impl U64Ext for u32 {
+    fn shift_by(&self, shift: u8, player: Player) -> u32 {
         match player {
             Player::South => self << shift,
             Player::North => self >> shift,
